Update seq_region_attrib queries
Remove top_level constraint, the first test this datacheck does is whether there is a single coord_system, thus by definition it will be top-level. Revert change that removed left-joined subquery - I think this was due to a misunderstanding of a previous comment; it will not work as expected to do two left joins instead, a nested subquery is needed.