Check that seq_region synonyms are not names of other seq_regions
(in the same coord system, for a given species)

If a seq_region synonym happens to be the name of a different
seq_region, then the API will prioritize the synonym over the name of
the seq_region, and so will be using the wrong sequence.

This happened in Phlebotomus papatasi, where the name of the scaffolds
were from a version of the assembly prior to its submission to INSDC
(the sequences ended up identical, but with different names in the final
version, hence the clash).