bugfix: 'multi' alias may not be available
When the registry includes the method
`Bio::EnsEMBL::Registry->load_registry_from_url()`, several aliases are
loaded by default (including 'multi', it seems). If only specific cores
are loaded using `Bio::EnsEMBL::DBSQL::DBAdaptor->new()`, this alias is
not added, making the DC fail. This is because `get_DBAdaptor()`, called
from `get_dba()`, throws an error when the species/alias passed as an
argument does not exist. Checking if it is there first fixes the issue,
preserving the previous behaviour.