Set SpeciesCommonName datacheck to be advisory.
The configuration requirements for this style of datacheck
turn out to be too onerous - it is working across multiple
core databases, and being able to locate all the relevant
dbs is a pain. This requires further thought as to how it
can best be integrated into our release processes, but for
now, set it as advisory, so that it is not a blocker.