Improving check for INSDC names/synonyms.
In order to embded the datacheck in a pipeline for exporting
GFF3 formatted for ENA submission, split the existing check
on INSDC names into two, because part of it is not relevant.
Tweak the INDSC name check to be a bit more general, and to
match the ENA submission use case, which requires all top-level
sequences to have an INSDC name/synonym.