Add new critical check for total ontology terms.
The namespaced counts might change from release to release,
as terms are moved around branches - but the totals for the
complete ontology should not, because terms are made obsolete
rather than deleted. So a new critical datacheck compares the
totals with last release.

Also tweaked the datacheck for required terms, so that if the
check fails, the message indicates which term is absent.