Apply karyotype_rank cardinality test to all seq_regions.
Chromosomes are often not indicated as such in the coord_system
table, with the current use of 'primary_assembly'. We want
to check for multiple 'karyotype_rank' attributes in both cases,
so remove the skipping element, and have a single SQL query
to test across all seq_regions at once.