Check format of 'genebuild.start_date'.
Non-ASCII characters in the meta_value of start_date meta_keys
cause compara to break (this field is loaded into genome_db
tables in compara dbs).
Also, check if start_date is the same as version - it shouldn't
be, that indicates that version is not being used for its
intended purpose.