Check for mitochondrial attribute.
We expect MT seq_regions to have a 'sequence_location' attrib,
with a value of 'mitochondrial_chromosome'. This was only being
checked if the coord_system name was 'chromosome' - but most
new core dbs do not use that name, they have 'primary_assembly'
instead.

This update splits out the MT stuff to a new datacheck, so that it
does not depend on the coord_system name.