Enforce INSDC accessions as seq_region names or synonyms.
For new species, and most old ones, we have INSDC accessions
for all of our sequences - this is really useful for users.
Due to execeptional cases on the main Ensembl site, however,
we don't enforce INSDC accessions, and the relevant datacheck
was advisory.

For Rapid Release, however, this data should always be available,
and we are introducing processes which depend on this (ENA-format
GFF3 dumps, and the new metadata schema).

This PR makes the datacheck critical for RR, remaining advisory
for the main site.