Fix compara datacheck retrieval of core dbs.
The compara datachecks that needed to load core dbs,
in order to check data consistency, were not working
if both a registry and server_uri were provided.
(If only one was provided, then assuming that the core
dbs were on the given server, everything was fine.)
When registry and server_uri are given, the code
looks in the metadata db, to try and work out the name
of the requisite core db. This requires a division
name (because three species are in two divisions, and
we need to only return one of them at a time), but
division metadata is formatted differently in compara
dbs, compared to all the species-specific dbs. Once
we account for that formatting, everything works as
expected.